requirements.txt
certifi==2024.7.0
python-dotenv==1.0.0
